Some 2.3.2 catchups
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2624 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/CLucene/index/TermVectorWriter.cpp
@@ -132,7 +132,7 @@ const char* TermVectorsWriter::LUCENE_TVF_EXTENSION = ".tvf";
   void TermVectorsWriter::addAllDocVectors(Array<TermFreqVector*>& vectors){
 	openDocument();
 
-	for (int32_t i = 0; i < vectors.length; ++i) {
+	for (size_t i = 0; i < vectors.length; ++i) {
 		bool storePositionWithTermVector = false;
 		bool storeOffsetWithTermVector = false;
